Having a separate XStepTest + XStepRunTest is an artifact of the 2.0 reactor split.
